modules/localtion service/requirements.txt
kafka-python
grpcio-tools